+++ test/e2e/main_test.go
@@ -3,7 +3,7 @@ package e2e
 import (
 	"context"
 	"fmt"
-	"github.com/DataDog/helm-charts/test/e2e/pulumi_env"
+	"github.com/DataDog/datadog-agent/test/new-e2e/utils/infra"
 	"os"
 	"testing"
 )
@@ -12,7 +12,7 @@ func TestMain(m *testing.M) {
 	code := m.Run()
 	if code == 0 {
 		fmt.Fprintf(os.Stderr, "Cleaning up stacks")
-		errs := pulumi_env.GetStackManager().Cleanup(context.Background())
+		errs := infra.GetStackManager().Cleanup(context.Background())
 		for _, err := range errs {
 			fmt.Fprint(os.Stderr, err.Error())
 		}
